moved banned words into its own dedicated package
