+++ src/banned-words/index.js
